requirement.txt
appdirs==1.4.4
astroid==2.4.2
cycler==0.10.0
distlib==0.3.1
filelock==3.0.12
isort==5.7.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.2.1
Pillow==8.1.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
virtualenv==20.4.0
wrapt==1.12.1
